Prepare populate database action

+++ app/src/main/java/com/tregula/graedukacyjna/domain/database/entity/CountryEntity.java
@@ -12,15 +12,15 @@ import androidx.room.PrimaryKey;
 public class CountryEntity {
 
     @PrimaryKey(autoGenerate = true)
-    public Integer id;
+    public Long id;
     public final String name;
     public final String location;
     public final String capitolName;
     public final String capitolLocation;
     public final String wiki;
-    public final Integer continentId;
+    public final Long continentId;
 
-    public CountryEntity(String name, String location, String capitolName, String capitolLocation, String wiki, Integer continentId) {
+    public CountryEntity(String name, String location, String capitolName, String capitolLocation, String wiki, Long continentId) {
         this.name = name;
         this.location = location;
         this.capitolName = capitolName;
